move file out of folder
